Post entity Presentation extension completed.
At least for now. On to the next thing.

RSpec: 732 examples, 0 failures, 1 pending
RuboCop: 240 files inspected, no offenses detected
